refactor: 에타 페이지 리팩토링
